m2_l5_simple-wiring-and-injection/src/test/java/com/ahanchar/service/impl/ProjectServiceImplTest.java
package com.ahanchar.service.impl;

import com.ahanchar.persistence.model.Project;
import com.ahanchar.persistence.repository.impl.ProjectRepositoryImpl;
import org.junit.Test;

import java.time.LocalDate;

import static org.junit.Assert.assertNotNull;

public class ProjectServiceImplTest {

    private ProjectServiceImpl projectServiceImpl = new ProjectServiceImpl(new ProjectRepositoryImpl());

    @Test
    public void givenNewProject_thenSavedSuccess() {
        Project newProject = new Project("First Project", LocalDate.now());

        assertNotNull(projectServiceImpl.save(newProject));
    }
}
